requirements.txt
boto3==1.34.0
google-cloud-bigquery==3.14.1
db-dtypes==1.2.0
pandas==2.1.4
pyyaml==6.0.1
python-dotenv==1.0.0
pytest==7.4.3
